Increase size of inputs that can be given
Inputs of 10000000000000000000 (B, KB, MB, GB, PB, EB) can be given.
However giving inputs of more than 18446744073709551615 cannot be properly
dealt with since it will overflow the unsigned long long int data type.
This will be fixed in an upcoming version of bytesize with the use of
the GMP library which is capable of handling bigger numbers.